se cambio el parseInt de la caja de porcentaje por un parseFloat
git-svn-id: https://repositorio.cidesa.com.ve/repositorio/web/php/siga/branches/productions/trunk@41852 bd7ca8a6-7b2b-0410-91d3-d9bcf4a077a3

+++ apps/nomina/modules/nomaumsuecar/templates/_gridcargos.php
@@ -27,7 +27,7 @@
 		{
 			if(parseInt(val))
 			{
-				$('ax_'+i+'_3').value=number_format(parseFloat(celda)+((parseFloat(celda)*parseInt(val))/100),2,',','.');	
+				$('ax_'+i+'_3').value=number_format(parseFloat(celda)+((parseFloat(celda)*parseFloat(val))/100),2,',','.');
 			}				
 			else
 			    break;
